requirements.txt
tensorflow>=2.20.0
numpy>=2.0.0
matplotlib
scikit-learn
pillow
streamlit
seaborn
